# cascade-oj/openapi.yaml
openapi: 3.0.3
info:
  title: AuthService API
  version: 0.0.1
...
paths:
...
  /aiops/v1/analysis/trigger:
    post:
      tags:
        - AIOps
      description: >-
        TriggerAnalysis manually triggers an alert analysis and noise reduction
        run.
      operationId: AIOps_TriggerAnalysis
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TriggerAnalysisRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TriggerAnalysisReply'
        default:
          description: Default error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
...
  /aiops/v1/reports:
    get:
      tags:
        - AIOps
      description: ListReports lists historical analysis reports.
      operationId: AIOps_ListReports
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            format: int32
        - name: pageSize
          in: query
          schema:
            type: integer
            format: int32
        - name: statusFilter
          in: query
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListReportsReply'
        default:
          description: Default error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
  '/aiops/v1/reports/{reportId}':
    get:
      tags:
        - AIOps
      description: GetReport retrieves a historical analysis report by ID or time range.
      operationId: AIOps_GetReport
      parameters:
        - name: reportId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetReportReply'
        default:
          description: Default error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
  '/aiops/v1/reports/{reportId}/ack':
    post:
      tags:
        - AIOps
      description: AckReport acknowledges a report.
      operationId: AIOps_AckReport
      parameters:
        - name: reportId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AckReportRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AckReportReply'
        default:
          description: Default error response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Status'
components:
  schemas:
    GoogleProtobufAny:
      type: object
      properties:
        '@type':
          type: string
          description: The type of the serialized message.
      additionalProperties: true
      description: >-
        Contains an arbitrary serialized message along with a @type that
        describes the type of the serialized message.
...
    Status:
      type: object
      properties:
        code:
          type: integer
          description: >-
            The status code, which should be an enum value of
            [google.rpc.Code][google.rpc.Code].
          format: int32
        message:
          type: string
          description: >-
            A developer-facing error message, which should be in English. Any
            user-facing error message should be localized and sent in the
            [google.rpc.Status.details][google.rpc.Status.details] field, or
            localized by the client.
        details:
          type: array
          items:
            $ref: '#/components/schemas/GoogleProtobufAny'
          description: >-
            A list of messages that carry the error details.  There is a common
            set of message types for APIs to use.
      description: >-
        The `Status` type defines a logical error model that is suitable for
        different programming environments, including REST APIs and RPC APIs. It
        is used by [gRPC](https://github.com/grpc). Each `Status` message
        contains three pieces of data: error code, error message, and error
        details. You can find out more about this error model and how to work
        with it in the [API Design
        Guide](https://cloud.google.com/apis/design/errors).
...
    AckReportReply:
      type: object
      properties: {}
    AckReportRequest:
      type: object
      properties:
        reportId:
          type: string
...
    GetReportReply:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        summary:
          type: string
        rootCausesJson:
          type: string
        suggestedActionsJson:
          type: string
        noiseCount:
          type: integer
          format: int32
        realRiskCount:
          type: integer
          format: int32
        status:
          type: string
        fullReportMarkdown:
          type: string
        createdAt:
          type: string
          format: date-time
        acknowledgedAt:
          type: string
          format: date-time
      description: GetReportReply contains the full report details.
    ListReportsReply:
      type: object
      properties:
        reports:
          type: array
          items:
            $ref: '#/components/schemas/ReportSummary'
        total:
          type: integer
          format: int32
      description: ListReportsReply contains a paginated list of report summaries.
    ReportSummary:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        noiseCount:
          type: integer
          format: int32
        realRiskCount:
          type: integer
          format: int32
        status:
          type: string
        createdAt:
          type: string
          format: date-time
      description: ReportSummary is a brief report overview for list display.
    TriggerAnalysisReply:
      type: object
      properties:
        reportId:
          type: string
        title:
          type: string
        summary:
          type: string
        noiseCount:
          type: integer
          format: int32
        realRiskCount:
          type: integer
          format: int32
        fullReportMarkdown:
          type: string
        createdAt:
          type: string
          format: date-time
      description: TriggerAnalysisReply contains the analysis result.
    TriggerAnalysisRequest:
      type: object
      properties:
        timeRangeMinutes:
          type: integer
          description: Optional time range in minutes before now (default 15).
          format: int32
        services:
          type: array
          items:
            type: string
          description: Optional list of service names to analyze (empty = all).
      description: >-
        TriggerAnalysisRequest is the request for manually triggering an
        analysis.